fix(memory): fix co-located INT4-vs-i64 decode mismatches surfaced by CI
CI's Postgres integration job caught a ColumnDecode failure on
admission_training.rs::get_training_batch (was_admitted/was_recalled are
INTEGER on Postgres, decoded as i64). A follow-up audit of every INTEGER
column against every site touched by the TIMESTAMPTZ sweep found the same
mismatch in 7 more places: persona.rs (evidence_count), preferences.rs
(evidence_count), session_digest.rs (token_count), compression_guidelines.rs
(version), skills.rs (invocation_count, version, success_count,
failure_count, plus an is_active BOOLEAN-vs-i64 mismatch in the same
tuple), messages/mod.rs::get_eviction_candidates (access_count), trust.rs
(requires_trust_check), and graph/store/mod.rs's EdgeRow (superseded_by,
turn_index).

Each site now decodes the INTEGER/BOOLEAN column at its correct type and
widens back to the existing public field type where one is exposed, so no
downstream signature changes.

Two pre-existing instances of the same class outside this PR's touched
functions (compression_guidelines.rs::load_compression_guidelines(_by_category)
and messages/mod.rs::find_promotion_candidates) were left untouched, since
this PR's diff never modifies those queries — tracked for a follow-up issue.

+++ crates/zeph-memory/src/store/admission_training.rs
@@ -154,6 +154,8 @@ impl SqliteStore {
         // `composite_score` is `REAL` (`FLOAT4`) on Postgres but the tuple decodes it as `f64`;
         // `CAST(... AS DOUBLE PRECISION)` widens it (same un-gated idiom already used for
         // `weight`/`confidence_fast`/`confidence_slow` in `graph/store/mod.rs`).
+        // `was_admitted`/`was_recalled` are `INTEGER` (`INT4`) on Postgres, so they decode as
+        // `i32`, not `i64` (INT8) — same rule as any other INTEGER/INT4 column in this crate.
         let created_at_sel =
             <zeph_db::ActiveDialect as zeph_db::dialect::Dialect>::select_as_text("created_at");
         let raw = format!(
@@ -174,8 +176,8 @@ impl SqliteStore {
                 String,
                 String,
                 f64,
-                i64,
-                i64,
+                i32,
+                i32,
                 String,
                 String,
             ),
